+++ source/torrent_source.cpp
@@ -145,7 +145,7 @@ bool torrent_source::open(void* ctx)
 	return true;
 }
 
-bool torrent_source::read_data(char* data, uint64_t offset, uint64_t size, uint64_t& read_size)
+bool torrent_source::read_data(char* data, uint64_t offset, size_t size, size_t& read_size)
 {
 	if (!m_read_op || !data || m_videos.size() == 0)
 		return false;
